fix: fixes some import issues caused by renaming
