+++ theoriq/dialog/__init__.py
@@ -8,8 +8,9 @@ from .router import RouteItem, RouterItemBlock
 from .runtime_error import ErrorItem, ErrorItemBlock
 from .text import TextItem, TextItemBlock
 from .web3 import Web3Item, Web3ItemBlock
-from .web3.eth.web3_eth_sign import Web3EthSignItem, Web3EthSignBlock
-from .web3.eth.web3_eth_sign_typed_data import (
+from .web3_blocks.eth.web3_eth_base import Web3EthBaseBlock
+from .web3_blocks.eth.web3_eth_sign import Web3EthSignItem, Web3EthSignBlock
+from .web3_blocks.eth.web3_eth_sign_typed_data import (
     Web3EthSignTypedDataItem,
     Web3EthSignTypedDataBlock,
     Web3EthTypedDataMessageType,
