+++ theoriq/dialog/web3_blocks/eth/web3_eth_sign.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, Literal, Optional
 
 from pydantic import BaseModel
 
-from .. import Web3Item
+from ...web3 import Web3Item
 from .web3_eth_base import Web3EthBaseBlock
